[SP] Fix 64 bit crash regarding collision map

+++ code/qcommon/cm_local.h
@@ -44,10 +44,10 @@ typedef struct {
 	int			cluster;
 	int			area;
 
-	int			firstLeafBrush;
+	intptr_t	firstLeafBrush;
 	int			numLeafBrushes;
 
-	int			firstLeafSurface;
+	intptr_t	firstLeafSurface;
 	int			numLeafSurfaces;
 } cLeaf_t;
